create JPA entities from DB, JPA findAll test

+++ cohort8th/rland-full/rland-boot3/src/main/java/kr/co/rland/boot3/repository/MenuRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Mapper
 public interface MenuRepository {
     //@Select("SELECT * FROM MENU")
-    List<MenuView> findAllByCategoryIdAndQuery(Integer offset, Integer size, List<Integer> categoryIds, String query);
+    List<MenuView> findAllByCategoryIdAndQuery(Integer offset, Integer size, List<Long> categoryIds, String query);
 //    List<MenuView> findAllWithImages(Integer categoryId, String query);
 
     void save(Menu menu);
